server/requirements.txt
langchain-ollama>=1.0.1
langgraph>=1.0.10
langgraph-checkpoint-postgres>=3.0.4
psycopg[binary]>=3.3.3
langchain-mcp-adapters>=0.2.1
python-dotenv>=1.2.2